requirements.txt.txt
opencv-python
face_recognition
psycopg2
Pillow